+++ src/03-posts/controllers/postsController.ts
@@ -10,8 +10,12 @@ export const getPostsController = async (req: Request,
     return res.status(200).json(posts)
 }
 
-export const findPostController = async (req: Request<{id: string}>,
+export const findPostController = async (req: Request<{ id: string }>,
                                          res: Response<PostType | {}>) => {
+    if (!ObjectId.isValid(req.params.id)) {
+        res.status(404).send()
+        return
+    }
     const foundPost = await PostsMongoDBRepository.findById(req.params.id)
     if (foundPost) {
         res.status(200).json(foundPost)
@@ -20,23 +24,31 @@ export const findPostController = async (req: Request<{id: string}>,
     }
 }
 
-export const createPostController = async (req: Request<{id: string}, {}, InputPostType>,
+export const createPostController = async (req: Request<{ id: string }, {}, InputPostType>,
                                            res: Response<PostType | Object>) => {
-     const blog = await BlogsMongoDBRepository.findById(req.body.blogId)
-     if (!blog) {
-         res.status(404).send('blog does not exist')
-         return
-     }
-     const newPost = await PostsMongoDBRepository.create(req.body, blog!.name)
-     if (!newPost) {
-         res.status(400).json()
-     } else {
-         res.status(201).json(newPost)
-     }
- }
+    if (!ObjectId.isValid(req.body.blogId)) {
+        res.status(404).send('blog does not exist')
+        return
+    }
+    const blog = await BlogsMongoDBRepository.findById(req.body.blogId)
+    if (!blog) {
+        res.status(404).send('blog does not exist')
+        return
+    }
+    const newPost = await PostsMongoDBRepository.create(req.body, blog!.name)
+    if (!newPost) {
+        res.status(400).json()
+    } else {
+        res.status(201).json(newPost)
+    }
+}
 
-export const updatePostController = async (req: Request<{id: string}, {}, InputPostType>,
+export const updatePostController = async (req: Request<{ id: string }, {}, InputPostType>,
                                            res: Response<PostType>) => {
+    if (!ObjectId.isValid(req.params.id)) {
+        res.status(404).send()
+        return
+    }
     const updatedPost = await PostsMongoDBRepository.update(req.params.id, req.body)
     if (updatedPost) {
         res.sendStatus(204)
@@ -45,12 +57,16 @@ export const updatePostController = async (req: Request<{id: string}, {}, InputP
     }
 }
 
-export const deletePostController = async (req: Request<{id: string}>,
+export const deletePostController = async (req: Request<{ id: string }>,
                                            res: Response) => {
+    if (!ObjectId.isValid(req.params.id)) {
+        res.status(404).send()
+        return
+    }
     const isDelete = await PostsMongoDBRepository.delete(req.params.id)
     if (isDelete) {
-        res.send(204)
+        res.sendStatus(204)
     } else {
-        res.send(404)
+        res.sendStatus(404)
     }
 }
